EVG-15258: move pod provisioning to cloud package
